Respond to more peer review
